requirements.txt
asgiref==3.7.2
click==8.1.7
Django==5.0.3
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-simple-history==3.5.0
drf-spectacular==0.27.1
gunicorn==21.2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
